# Optimize ser/de to avoid using output stream